trouble with gmails spam filter
Not sure if this is a common problem. But all emails i was sending to my gmail account kept ending up in the spam folder.

Adding this seems to of fixed it for ME. 

However I believe it will show all email addresses in the TO: field now.

Might need someone else to test this.