Add a very basic Windows notification service. Spawns an icon and "balloontip" in the notification area. There are limitations to this method (see the powershell script file), but it meets my needs.
Python script that calls the powershell script using python's subprocess module.
There may be a better way to do this, but it works so far!

Includes an MQTT icon from mqtt.org that is used for the notification area
icon.
Includes two example screenshots of the end result while using - an ideal, and an example of what
happens if there are too many messages (lots of icons you can not read the message from).