Support urls in osxnotify
When using osx desktop notification, it can be useful to have a
URL associated to the notification, so that it may be clicked to
open a web page with further details.

I'm using this with Firehose from OpenStack when listening to gerrit
topics, to open the corresponding change in gerrit. The url can be
set un item.data using payload transformation functions, e.g.

def all_data(topic, data, srv):
    extra_data = json.loads(data['payload'])
    # Make the change data (including url) available for formatting
    return extra_data['change']