build post data correctly when no additional influx tags included
It currently seems assumes that in line 55 by appending `item.message`  after a comma, any message includes at least one more influx tag to start with. This may not always be the case. Enhancing the logic to build the data payload for Influx post in a way that format strings can work without having to add additional tags.