Tests: Improve mocking of the `socket` object using `pytest-mock`
In this way, it is made sure that all mocks are cleared up after the
exit of the test case function.

The flaw was revealed when the `carbon` and `irccat` test cases, which
apparently left some mocks on the `socket` object behind, were combined
with the new `e2e` test cases, which in turn use the HTTP protocol to
communicate with the Docker daemon.