Fix notify_only_on_timeout behavior
If notify_only_on_timeout is set, mqttwarn will send a notification in case the timeout was reached.
With this change, it will also send a message in case a message has been received within the timeout for the first time.
The assumption is, that only timeout/no timeout is relevant and if messages are received within the timeout, that's not a status change.
If all messages are of relevant, just set notify_only_on_timeout = False which is the default.